TreeGUISvc -restart, fixed TreeGUISvc indexing
Resetting the interval affects when the Elapsed event is raised. For
example, if you set the interval to 5 seconds and then set the Enabled
property to true, the count starts at the time Enabled is set. If you
reset the interval to 10 seconds when the count is 3 seconds, the
Elapsed event is raised for the first time 13 seconds after the Enabled
property was set to true.